requirement.txt
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.1
cssselect2==0.8.0
docx2pdf==0.1.8
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.59.0
html5lib==1.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.2.2
pillow==11.3.0
protobuf==4.21.12
pycparser==2.22
pydyf==0.11.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyMySQL==1.1.0
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2025.2
reportlab==4.0.4
six==1.17.0
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
weasyprint==60.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
